Extract memcached configs from MenuCache service into its own MemcachedClient service - dependency injection
