+++ packages/shift-next-routes/src/lib/menu-cache.js
@@ -1,20 +1,5 @@
-// Libraries
-const Memcached = require('memcache-client')
-
-// Config
-const memcachedServers = (process.env.MEMCACHED_SERVERS || '').split(',')
-const memcachedConfig = {
-  server: {
-    servers: memcachedServers.map((server) => {
-      return { server: server, maxConnections: 5 }
-    }),
-    config: {
-      retryFailedServerInterval: 1000, // (ms) - how often to check failed servers
-      failedServerOutTime: 30000 // (ms) how long a failed server should be out before retrying it
-    }
-  },
-  cmdTimeout: 1500 // (ms) command timeout
-}
+// Libs
+const { MemcachedClient } = require('./memcached-client')
 
 /**
  *  MenuCache service
@@ -29,7 +14,7 @@ class MenuCache {
    * @param {function} cache - cache client, eg. memcached
    * @param {string} cacheKey - the cache key to be used
    */
-  constructor (cache = new Memcached(memcachedConfig), cacheKey = 'menus/data') {
+  constructor (cache = MemcachedClient, cacheKey = 'menus/data') {
     this.cache = cache
     this.cacheKey = cacheKey
   }
